Implement `sys_ppoll` for crate `poll`
